refactor: isolate init code specific to workflow runs
